requirements.txt
aniso8601==9.0.1
blinker==1.8.2
chardet==5.2.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
pillow==11.0.0
pytz==2024.2
reportlab==4.2.5
six==1.16.0
Werkzeug==3.0.4
WTForms==3.1.2
